Remove redundant indexes in MySQL DB schema
- Removes unnecessary indexes: InnoDB includes the primary key in all secondary indexes.  A few of
the indexes defined in this schema were either a subset of a duplicate of the primary key.

- Adds explicit primary key to `zipkin_annotations` to avoid duplicate GEN_CLUST_INDEX:  According to
the MySQL docs, if a primary key is not defined "MySQL locates the first UNIQUE index where all the key
columns are NOT NULL and InnoDB uses it as the clustered index."  `zipkin_annotations` defines a
unique index however the index contains a NULL column, so MySQL does NOT use it as the primary key.
As a result, a GEN_CLUST_INDEX is created alongside the UNIQUE index.  The redundant index can be
eliminated if the unique index is instead defined as a primary key.